Continued refactoring of the home screen, as well as a simplification of the drawer on the home screen, now called the profile drawer, as it is only for accessing the profile now. Home screen now includes all information, like about us and john cage, rather than getting to it from the drawer. Can now traverse the different screens by swiping rather than the nav bar.
